generalize app state a bit, move to a shared location
